Use finally
git-svn-id: svn://svn.code.sf.net/p/jamon/code/trunk/jamon@2226 6afd1a24-8710-0410-9bc4-c69fca332c4c

+++ core/src/org/jamon/TemplateProcessor.java
@@ -67,23 +67,20 @@ public class TemplateProcessor
         throws IOException
     {
         FileOutputStream out = new FileOutputStream(javaFile);
+        boolean success = false;
         try
         {
             sourceGenerator.generateSource(out);
+            success = true;
         }
-        catch (RuntimeException e)
+        finally
         {
             out.close();
-            javaFile.delete();
-            throw e;
+            if (!success)
+            {
+                javaFile.delete();
+            }
         }
-        catch (IOException e)
-        {
-            out.close();
-            javaFile.delete();
-            throw e;
-        }
-        out.close();
     }
 
     private static void showHelp()
